modelviewset 수정/삭제 권한 조건 변경

+++ memos/views.py
@@ -1,13 +1,14 @@
 from rest_framework import status, viewsets
 from rest_framework.response import Response
 from rest_framework.exceptions import NotFound
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
+# from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from conf.permissions import ReadOnlyIfNotAuthenticatedOrOwner
 from .models import Memo
 from .serializers import MemoSerializer
 
 
 class MemoViewSet(viewsets.ModelViewSet):
-  permission_classes = [IsAuthenticatedOrReadOnly]
+  permission_classes = [ReadOnlyIfNotAuthenticatedOrOwner]
   serializer_class = MemoSerializer
   queryset = Memo.objects.prefetch_related('user').all()
 
@@ -46,6 +47,7 @@ class MemoViewSet(viewsets.ModelViewSet):
     
   def update(self, request, pk=None, *args, **kwargs):
     memo = self.get_object(pk)
+    self.check_object_permissions(self.request, memo)
     serializer = self.serializer_class(
       memo,
       data=request.data,
@@ -59,6 +61,7 @@ class MemoViewSet(viewsets.ModelViewSet):
     
   def destroy(self, request, pk=None, *args, **kwargs):
     memo = self.get_object(pk)
+    self.check_object_permissions(self.request, memo)
     memo.delete()
 
     return Response(status=status.HTTP_204_NO_CONTENT)
